Implemented and improved the new Watcher implementation test

+++ CmisSync/TestLibrary/SyncStrategiesTests/WatcherTest.cs
@@ -1,5 +1,6 @@
 using System;
 using System.IO;
+using System.Threading.Tasks;
 
 using CmisSync.Lib.Events;
 using CmisSync.Lib.Sync.Strategy;
@@ -14,25 +15,35 @@ namespace TestLibrary.SyncStrategiesTests
     public class WatcherTest
     {
 
-        private string localPath; 
+        private string localPath;
         private DirectoryInfo localFolder;
         private FileInfo localFile;
+        private DirectoryInfo localSubFolder;
+        private static readonly int RETRIES = 500;
+        private static readonly int MILISECONDSWAIT = 100;
 
         [SetUp]
         public void SetUp() {
             localPath = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName());
             localFolder = new DirectoryInfo(localPath);
             localFolder.Create();
+            localSubFolder = new DirectoryInfo(Path.Combine(localFolder.FullName, Path.GetRandomFileName()));
+            localSubFolder.Create();
             localFile = new FileInfo(Path.Combine(localFolder.FullName, Path.GetRandomFileName()));
-            localFile.Create();
+            using(localFile.Create());
         }
 
         [TearDown]
         public void TearDown() {
             localFile.Refresh();
-            localFile.Delete();
+            if(localFile.Exists)
+                localFile.Delete();
+            localSubFolder.Refresh();
+            if(localSubFolder.Exists)
+                localSubFolder.Delete(true);
             localFolder.Refresh();
-            localFolder.Delete();
+            if(localFolder.Exists)
+                localFolder.Delete(true);
         }
 
         [Test, Category("Fast")]
@@ -217,117 +228,218 @@ namespace TestLibrary.SyncStrategiesTests
         }
 
 
-        [Ignore]
-        [Test, Category("Fast")]
+        [Test, Category("Medium")]
         public void ReportFSFileAddedEventTest () {
-            var fswatcher = new Mock<FileSystemWatcher>(localFolder.FullName){CallBase = false};
+            localFile.Delete();
+            localFile = new FileInfo(Path.Combine(localFolder.FullName, Path.GetRandomFileName()));
+            var fswatcher = new FileSystemWatcher(localFolder.FullName);
             var manager = new Mock<SyncEventManager>().Object;
             var queue = new Mock<SyncEventQueue>(manager);
             FSEvent returnedFSEvent = null;
             queue.Setup(q => q.AddEvent(It.IsAny<FSEvent>()))
                 .Callback((ISyncEvent file) => returnedFSEvent = file as FSEvent);
-            var watcher = new Watcher(fswatcher.Object, queue.Object);
-            fswatcher.Raise(m => m.Created += null, null, new FileSystemEventArgs(WatcherChangeTypes.Created,localFolder.FullName, localFile.Name));
+            var watcher = new Watcher(fswatcher, queue.Object);
+            watcher.EnableEvents = true;
+            var t = Task.Factory.StartNew(() => {
+                int count = 0;
+                while(returnedFSEvent == null && count < RETRIES) {
+                    fswatcher.WaitForChanged(WatcherChangeTypes.Created, MILISECONDSWAIT);
+                    count ++;
+                }
+            });
+            using(localFile.Create());
+            t.Wait();
+            Assert.NotNull(returnedFSEvent);
+            Assert.IsFalse(returnedFSEvent.IsDirectory());
             Assert.AreEqual(localFile.FullName, returnedFSEvent.Path);
             Assert.AreEqual(WatcherChangeTypes.Created, returnedFSEvent.Type);
         }
 
-        [Ignore]
-        [Test, Category("Fast")]
+        [Test, Category("Medium")]
         public void ReportFSFileChangedEventTest () {
-            var fswatcher = new Mock<FileSystemWatcher>(localFolder.FullName){CallBase = true}.Object;
+            var fswatcher = new FileSystemWatcher(localFolder.FullName);
             var manager = new Mock<SyncEventManager>().Object;
             var queue = new Mock<SyncEventQueue>(manager);
             FSEvent returnedFSEvent = null;
-            queue.Setup(q => q.AddEvent(It.IsAny<AbstractFolderEvent>()))
+            queue.Setup(q => q.AddEvent(It.IsAny<FSEvent>()))
                 .Callback((ISyncEvent file) => returnedFSEvent = file as FSEvent);
             var watcher = new Watcher(fswatcher, queue.Object);
-            
-            Assert.Fail ();
+            watcher.EnableEvents = true;
+            var t = Task.Factory.StartNew(() => {
+                int count = 0;
+                while(returnedFSEvent == null && count < RETRIES) {
+                    fswatcher.WaitForChanged(WatcherChangeTypes.Changed, MILISECONDSWAIT);
+                    count ++;
+                }
+            });
+            using (FileStream stream = File.OpenWrite(localFile.FullName))
+            {
+                byte[] data = new byte[1024];
+                // Write data
+                stream.Write(data, 0, data.Length);
+            }
+            t.Wait();
+            Assert.NotNull(returnedFSEvent);
+            Assert.IsFalse(returnedFSEvent.IsDirectory());
+            Assert.AreEqual(localFile.FullName, returnedFSEvent.Path);
+            Assert.AreEqual(WatcherChangeTypes.Changed, returnedFSEvent.Type);
         }
 
-        [Ignore]
-        [Test, Category("Fast")]
+        [Test, Category("Medium")]
         public void ReportFSFileRenamedEventTest () {
-            var fswatcher = new Mock<FileSystemWatcher>(localFolder.FullName){CallBase = true}.Object;
+            var fswatcher = new FileSystemWatcher(localFolder.FullName);
             var manager = new Mock<SyncEventManager>().Object;
             var queue = new Mock<SyncEventQueue>(manager);
+            string oldpath = localFile.FullName;
+            string newpath = Path.Combine(localFolder.FullName, Path.GetRandomFileName());
             FSEvent returnedFSEvent = null;
-            queue.Setup(q => q.AddEvent(It.IsAny<AbstractFolderEvent>()))
+            queue.Setup(q => q.AddEvent(It.IsAny<FSEvent>()))
                 .Callback((ISyncEvent file) => returnedFSEvent = file as FSEvent);
             var watcher = new Watcher(fswatcher, queue.Object);
-            
-            Assert.Fail ();
+            watcher.EnableEvents = true;
+            var t = Task.Factory.StartNew(() => {
+                int count = 0;
+                while(returnedFSEvent == null && count < RETRIES) {
+                    fswatcher.WaitForChanged(WatcherChangeTypes.Renamed, MILISECONDSWAIT);
+                    count ++;
+                }
+            });
+            localFile.MoveTo(newpath);
+            t.Wait();
+            Assert.NotNull(returnedFSEvent);
+            Assert.IsFalse(returnedFSEvent.IsDirectory());
+            Assert.AreEqual(newpath, (returnedFSEvent as FSMovedEvent).Path);
+            Assert.AreEqual(oldpath, (returnedFSEvent as FSMovedEvent).OldPath);
+            Assert.AreEqual(WatcherChangeTypes.Renamed, (returnedFSEvent as FSMovedEvent).Type);
+            localFile = new FileInfo(newpath);
         }
 
-        [Ignore]
-        [Test, Category("Fast")]
+        [Test, Category("Medium")]
         public void ReportFSFileRemovedEventTest () {
-            var fswatcher = new Mock<FileSystemWatcher>(localFolder.FullName){CallBase = true}.Object;
+            var fswatcher = new FileSystemWatcher(localFolder.FullName);
             var manager = new Mock<SyncEventManager>().Object;
             var queue = new Mock<SyncEventQueue>(manager);
             FSEvent returnedFSEvent = null;
-            queue.Setup(q => q.AddEvent(It.IsAny<AbstractFolderEvent>()))
+            queue.Setup(q => q.AddEvent(It.IsAny<FSEvent>()))
                 .Callback((ISyncEvent file) => returnedFSEvent = file as FSEvent);
             var watcher = new Watcher(fswatcher, queue.Object);
-            
-            Assert.Fail ();
+            watcher.EnableEvents = true;
+            var t = Task.Factory.StartNew(() => {
+                int count = 0;
+                while(returnedFSEvent == null && count < RETRIES) {
+                    fswatcher.WaitForChanged(WatcherChangeTypes.Deleted, MILISECONDSWAIT);
+                    count ++;
+                }
+            });
+            localFile.Delete();
+            t.Wait();
+            Assert.NotNull(returnedFSEvent);
+            Assert.AreEqual(localFile.FullName, returnedFSEvent.Path);
+            Assert.AreEqual(WatcherChangeTypes.Deleted, returnedFSEvent.Type);
         }
 
-        [Ignore]
-        [Test, Category("Fast")]
+        [Test, Category("Medium")]
         public void ReportFSFolderAddedEventTest () {
-            var fswatcher = new Mock<FileSystemWatcher>(localFolder.FullName){CallBase = true}.Object;
+            localSubFolder.Delete();
+            var fswatcher = new FileSystemWatcher(localFolder.FullName);
             var manager = new Mock<SyncEventManager>().Object;
             var queue = new Mock<SyncEventQueue>(manager);
             FSEvent returnedFSEvent = null;
-            queue.Setup(q => q.AddEvent(It.IsAny<AbstractFolderEvent>()))
+            queue.Setup(q => q.AddEvent(It.IsAny<FSEvent>()))
                 .Callback((ISyncEvent file) => returnedFSEvent = file as FSEvent);
             var watcher = new Watcher(fswatcher, queue.Object);
-            
-            Assert.Fail ();
+            watcher.EnableEvents = true;
+            var t = Task.Factory.StartNew(() => {
+                int count = 0;
+                while(returnedFSEvent == null && count < RETRIES) {
+                    fswatcher.WaitForChanged(WatcherChangeTypes.Created, MILISECONDSWAIT);
+                    count ++;
+                }
+            });
+            localSubFolder.Create();
+            t.Wait();
+            Assert.NotNull(returnedFSEvent);
+            Assert.IsTrue(returnedFSEvent.IsDirectory());
+            Assert.AreEqual(localSubFolder.FullName, returnedFSEvent.Path);
+            Assert.AreEqual(WatcherChangeTypes.Created, returnedFSEvent.Type);
         }
 
-        [Ignore]
-        [Test, Category("Fast")]
+        [Test, Category("Medium")]
         public void ReportFSFolderChangedEventTest () {
-            var fswatcher = new Mock<FileSystemWatcher>(localFolder.FullName){CallBase = true}.Object;
+            var fswatcher = new FileSystemWatcher(localFolder.FullName);
             var manager = new Mock<SyncEventManager>().Object;
             var queue = new Mock<SyncEventQueue>(manager);
             FSEvent returnedFSEvent = null;
-            queue.Setup(q => q.AddEvent(It.IsAny<AbstractFolderEvent>()))
+            queue.Setup(q => q.AddEvent(It.IsAny<FSEvent>()))
                 .Callback((ISyncEvent file) => returnedFSEvent = file as FSEvent);
             var watcher = new Watcher(fswatcher, queue.Object);
-            
-            Assert.Fail ();
+            watcher.EnableEvents = true;
+            var t = Task.Factory.StartNew(() => {
+                int count = 0;
+                while(returnedFSEvent == null && count < RETRIES) {
+                    fswatcher.WaitForChanged(WatcherChangeTypes.Changed, MILISECONDSWAIT);
+                    count ++;
+                }
+            });
+            localSubFolder.CreationTime = DateTime.Now;
+            t.Wait();
+            Assert.NotNull(returnedFSEvent);
+            Assert.IsTrue(returnedFSEvent.IsDirectory());
+            Assert.AreEqual(localSubFolder.FullName, returnedFSEvent.Path);
+            Assert.AreEqual(WatcherChangeTypes.Changed, returnedFSEvent.Type);
         }
 
-        [Ignore]
-        [Test, Category("Fast")]
+        [Test, Category("Medium")]
         public void ReportFSFolderRemovedEventTest () {
-            var fswatcher = new Mock<FileSystemWatcher>(localFolder.FullName){CallBase = true}.Object;
+            var fswatcher = new FileSystemWatcher(localFolder.FullName);
             var manager = new Mock<SyncEventManager>().Object;
             var queue = new Mock<SyncEventQueue>(manager);
             FSEvent returnedFSEvent = null;
-            queue.Setup(q => q.AddEvent(It.IsAny<AbstractFolderEvent>()))
+            queue.Setup(q => q.AddEvent(It.IsAny<FSEvent>()))
                 .Callback((ISyncEvent file) => returnedFSEvent = file as FSEvent);
             var watcher = new Watcher(fswatcher, queue.Object);
-            
-            Assert.Fail ();
+            watcher.EnableEvents = true;
+            var t = Task.Factory.StartNew(() => {
+                int count = 0;
+                while(returnedFSEvent == null && count < RETRIES) {
+                    fswatcher.WaitForChanged(WatcherChangeTypes.Deleted, MILISECONDSWAIT);
+                    count ++;
+                }
+            });
+            localSubFolder.Delete();
+            t.Wait();
+            Assert.NotNull(returnedFSEvent);
+            Assert.AreEqual(localSubFolder.FullName, returnedFSEvent.Path);
+            Assert.AreEqual(WatcherChangeTypes.Deleted, returnedFSEvent.Type);
         }
 
-        [Ignore]
-        [Test, Category("Fast")]
+        [Test, Category("Medium")]
         public void ReportFSFolderRenamedEventTest () {
-            var fswatcher = new Mock<FileSystemWatcher>(localFolder.FullName){CallBase = true}.Object;
+            var fswatcher = new FileSystemWatcher(localFolder.FullName);
             var manager = new Mock<SyncEventManager>().Object;
             var queue = new Mock<SyncEventQueue>(manager);
+            string oldpath = localSubFolder.FullName;
+            string newpath = Path.Combine(localFolder.FullName, Path.GetRandomFileName());
             FSEvent returnedFSEvent = null;
-            queue.Setup(q => q.AddEvent(It.IsAny<AbstractFolderEvent>()))
+            queue.Setup(q => q.AddEvent(It.IsAny<FSEvent>()))
                 .Callback((ISyncEvent file) => returnedFSEvent = file as FSEvent);
             var watcher = new Watcher(fswatcher, queue.Object);
-            
-            Assert.Fail ();
+            watcher.EnableEvents = true;
+            var t = Task.Factory.StartNew(() => {
+                int count = 0;
+                while(returnedFSEvent == null && count < RETRIES) {
+                    fswatcher.WaitForChanged(WatcherChangeTypes.Renamed, MILISECONDSWAIT);
+                    count ++;
+                }
+            });
+            localSubFolder.MoveTo(newpath);
+            t.Wait();
+            Assert.NotNull(returnedFSEvent);
+            Assert.IsTrue(returnedFSEvent.IsDirectory());
+            Assert.AreEqual(oldpath, (returnedFSEvent as FSMovedEvent).OldPath);
+            Assert.AreEqual(newpath, (returnedFSEvent as FSMovedEvent).Path);
+            Assert.AreEqual(WatcherChangeTypes.Renamed, returnedFSEvent.Type);
+            localSubFolder = new DirectoryInfo(newpath);
         }
     }
 }
